Fix missing GRUB on minimal systems
On systems like GitHub actions runners (and probably containers), there
is no GRUB, so the install script cannot read or modify it. While we
don't really want users to be installing this package in those
situations, if they do, we should fail as gracefully as possible while
warning them about the issue.

Signed-off-by: June Knauth (tt) <jknauth@tenstorrent.com>